Struggling with private structs.
